Подключил карту к странице и реализовал активацию страници при инициализации карты
