+++ js/main.js
@@ -1,25 +1,5 @@
 import {
-  activateMapFilters,
-  deactivateMapFilters
-} from './map-filters.js';
-
-import {
-  activateForm,
-  deactivateForm
-} from './form.js';
-
-import {
 } from './form-validation.js';
 
-const activatePage = () => {
-  activateMapFilters();
-  activateForm();
-};
-
-const deactivatePage = () => {
-  deactivateMapFilters();
-  deactivateForm();
-};
-
-deactivatePage();
-activatePage();
+import {
+} from './map.js';
