+++ js/map.js
@@ -1,0 +1,38 @@
+import {
+  activateMapFilters,
+  deactivateMapFilters
+} from './map-filters.js';
+
+import {
+  activateForm,
+  deactivateForm
+} from './form.js';
+
+
+const activatePage = () => {
+  activateMapFilters();
+  activateForm();
+};
+
+const deactivatePage = () => {
+  deactivateMapFilters();
+  deactivateForm();
+};
+
+deactivatePage();
+
+const map = L.map('map-canvas')
+  .on('load', () => {
+    activatePage();
+  })
+  .setView({
+    lat: 35.681700,
+    lng: 139.753891,
+  }, 10);
+
+L.tileLayer(
+  'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
+  {
+    attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
+  },
+).addTo(map);
